refactor: - changed CommandParsingContext class name to CommandParsingStrategiesManager - improved DIP by creating CommandParsingStrategiesManager constructor which takes List of ParsingStrategy with all strategies (TestJobs2dApp creates ready to go CommandParsingStrategiesManager) - removed unnecessary dependencies (importing CsvParsingStrategy and JsonParsingStrategy in CommandParsingStrategiesManager)
